Renamed copy-class-slots to slot-list-for-copy and added instance parameter so it can be specialized on the instance type not just metaclass type.
Extract actual copy of slots into new exported function copy-instance-slots

Add arbitrary instance-copy-reason parameter to copy-instance and
everything else called by it, which can be specialized on. For example
copying an object for backing when editing in a GUI may be different
from copying it to send it over the wire.